Extract Default Converter
