+++ FhirToDataLake/schemaTemplates/ExamplePatient.liquid
@@ -1,0 +1,6 @@
+{% validate "ExamplePatient.schema.json" -%}
+{
+    "resourceType": "{{ msg.resourceType }}",
+    "id": "{{ msg.id }}"
+}
+{% endvalidate -%}
