+++ FhirToDataLake/schemaTemplates/ExamplePatient.schema.json
@@ -1,0 +1,9 @@
+{
+  "title": "Patient customized schema",
+  "type": "object",
+  "properties": {
+    "resourceType": { "type": "string" },
+	"id": { "type": "string" }
+  },
+  "required": [ "id" ]
+}
