+++ FhirToDataLake/schemaTemplates/Patient.liquid
@@ -1,0 +1,4 @@
+{
+    "resourceType": "{{ msg.resourceType }}",
+    "id": "{{ msg.id }}"
+}
