+++ FhirToDataLake/schemaTemplates/metadata.json
@@ -1,0 +1,3 @@
+{
+	"type": "json"
+}
