+++ FhirToDataLake/src/Microsoft.Health.Fhir.Synapse.Core/DataProcessor/ParquetDataProcessor.cs
@@ -15,6 +15,7 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Microsoft.Health.Fhir.Synapse.Common.Configurations.Arrow;
 using Microsoft.Health.Fhir.Synapse.Common.Models.Data;
+using Microsoft.Health.Fhir.Synapse.Core.DataProcessor.DataConverter;
 using Microsoft.Health.Fhir.Synapse.Core.Exceptions;
 using Microsoft.Health.Fhir.Synapse.Parquet.CLR;
 using Microsoft.Health.Fhir.Synapse.SchemaManagement;
@@ -30,10 +31,14 @@ namespace Microsoft.Health.Fhir.Synapse.Core.DataProcessor
         private readonly ArrowConfiguration _arrowConfiguration;
         private readonly ILogger<ParquetDataProcessor> _logger;
         private readonly ParquetConverterWrapper _parquetConverterWrapper;
+        private readonly DefaultConverter _defaultDataConverter;
+        private readonly FhirConverter _fhirDataConverter;
 
         public ParquetDataProcessor(
             IFhirSchemaManager<FhirParquetSchemaNode> fhirSchemaManager,
             IOptions<ArrowConfiguration> arrowConfiguration,
+            DefaultConverter defaultConverter,
+            FhirConverter fhirConverter,
             ILogger<ParquetDataProcessor> logger)
         {
             EnsureArg.IsNotNull(fhirSchemaManager, nameof(fhirSchemaManager));
@@ -42,6 +47,8 @@ namespace Microsoft.Health.Fhir.Synapse.Core.DataProcessor
 
             _fhirSchemaManager = fhirSchemaManager;
             _arrowConfiguration = arrowConfiguration.Value;
+            _defaultDataConverter = defaultConverter;
+            _fhirDataConverter = fhirConverter;
             _logger = logger;
 
             try
@@ -64,9 +71,6 @@ namespace Microsoft.Health.Fhir.Synapse.Core.DataProcessor
         {
             cancellationToken.ThrowIfCancellationRequested();
 
-                // Preprocess data
-            JsonBatchData preprocessedData = Preprocess(inputData, processParameters.SchemaType, cancellationToken);
-
             // Get FHIR schema for the input data.
             var schema = _fhirSchemaManager.GetSchema(processParameters.SchemaType);
 
@@ -76,7 +80,10 @@ namespace Microsoft.Health.Fhir.Synapse.Core.DataProcessor
                 throw new ParquetDataProcessorException($"The FHIR schema node could not be found for schema type '{processParameters.SchemaType}'.");
             }
 
-            var inputStream = ConvertJsonDataToStream(processParameters.SchemaType, preprocessedData.Values);
+            // Convert data based on schema
+            JsonBatchData processedData = _defaultDataConverter.Convert(inputData, schema, cancellationToken);
+
+            var inputStream = TransformJsonDataToStream(processParameters.SchemaType, processedData.Values);
             if (inputStream == null)
             {
                 // Return null if no data has been converted.
@@ -90,9 +97,8 @@ namespace Microsoft.Health.Fhir.Synapse.Core.DataProcessor
                 return Task.FromResult(
                     new StreamBatchData(
                         resultStream,
-                        preprocessedData.Values.Count(),
-                        processParameters.SchemaType)
-                    );
+                        processedData.Values.Count(),
+                        processParameters.SchemaType));
             }
             catch (Exception ex)
             {
@@ -101,145 +107,7 @@ namespace Microsoft.Health.Fhir.Synapse.Core.DataProcessor
             }
         }
 
-        public JsonBatchData Preprocess(
-            JsonBatchData inputData,
-            string schemaType,
-            CancellationToken cancellationToken = default)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-
-            // Get FHIR schema for the input data.
-            var schema = _fhirSchemaManager.GetSchema(schemaType);
-
-            if (schema == null)
-            {
-                _logger.LogError($"The FHIR schema node could not be found for schema type '{schemaType}'.");
-                throw new ParquetDataProcessorException($"The FHIR schema node could not be found for schema type '{schemaType}'.");
-            }
-
-            var processedJsonData = inputData.Values
-                .Select(json => ProcessStructObject(json, schema))
-                .Where(processedResult => processedResult != null);
-
-            return new JsonBatchData(processedJsonData);
-        }
-
-        private JObject ProcessStructObject(JToken structItem, FhirParquetSchemaNode schemaNode)
-        {
-            if (structItem is not JObject fhirJObject)
-            {
-                _logger.LogError($"Current FHIR object is not a valid JObject: {schemaNode.GetNodePath()}.");
-                throw new ParquetDataProcessorException($"Current FHIR object is not a valid JObject: {schemaNode.GetNodePath()}.");
-            }
-
-            JObject processedObject = new JObject();
-
-            foreach (var subItem in fhirJObject)
-            {
-                JToken subObject = subItem.Value;
-
-                // Process choice type FHIR resource.
-                if (schemaNode.ContainsChoiceDataType(subItem.Key))
-                {
-                    var choiceTypeName = schemaNode.ChoiceTypeNodes[subItem.Key].Item1;
-                    var choiceTypeDataType = schemaNode.ChoiceTypeNodes[subItem.Key].Item2;
-
-                    if (!schemaNode.SubNodes[choiceTypeName].SubNodes.ContainsKey(choiceTypeDataType))
-                    {
-                        _logger.LogError($"Data type \"{choiceTypeDataType}\" cannot be found in choice type property, {schemaNode.GetNodePath()}.");
-                        throw new ParquetDataProcessorException($"Data type \"{choiceTypeDataType}\" cannot be found in choice type property, {schemaNode.GetNodePath()}.");
-                    }
-
-                    var dataTypeNode = schemaNode.SubNodes[choiceTypeName].SubNodes[choiceTypeDataType];
-                    processedObject.Add(choiceTypeName, ProcessChoiceTypeObject(subObject, dataTypeNode));
-                }
-                else
-                {
-                    // Ignore FHIR data node if it doesn't exist in schema.
-                    if (schemaNode.SubNodes == null || !schemaNode.SubNodes.ContainsKey(subItem.Key))
-                    {
-                        continue;
-                    }
-
-                    FhirParquetSchemaNode subNode = schemaNode.SubNodes[subItem.Key];
-
-                    if (subNode.IsRepeated)
-                    {
-                        // Process array FHIR resource.
-                        processedObject.Add(subNode.Name, ProcessArrayObject(subObject, subNode));
-                    }
-                    else if (subNode.IsLeaf)
-                    {
-                        // Process leaf FHIR resource.
-                        processedObject.Add(subNode.Name, ProcessLeafObject(subObject, subNode));
-                    }
-                    else
-                    {
-                        // Process struct FHIR resource.
-                        processedObject.Add(subNode.Name, ProcessStructObject(subObject, subNode));
-                    }
-                }
-            }
-
-            return processedObject;
-        }
-
-        private JArray ProcessArrayObject(JToken arrayItem, FhirParquetSchemaNode schemaNode)
-        {
-            if (arrayItem is not JArray fhirArrayObject)
-            {
-                _logger.LogError($"Current FHIR object is not a valid JArray: {schemaNode.GetNodePath()}.");
-                throw new ParquetDataProcessorException($"Current FHIR object is not a valid JArray: {schemaNode.GetNodePath()}.");
-            }
-
-            JArray arrayObject = new JArray();
-            foreach (var item in fhirArrayObject)
-            {
-                if (schemaNode.IsLeaf)
-                {
-                    arrayObject.Add(ProcessLeafObject(item, schemaNode));
-                }
-                else
-                {
-                    arrayObject.Add(ProcessStructObject(item, schemaNode));
-                }
-            }
-
-            return arrayObject;
-        }
-
-        private JValue ProcessLeafObject(JToken fhirObject, FhirParquetSchemaNode schemaNode)
-        {
-            if (schemaNode.Type == FhirParquetSchemaNodeConstants.JsonStringType)
-            {
-                return new JValue(fhirObject.ToString(Formatting.None));
-            }
-
-            if (fhirObject is not JValue fhirLeafObject)
-            {
-                _logger.LogError($"Invalid data: complex object found in leaf schema node {schemaNode.GetNodePath()}.");
-                throw new ParquetDataProcessorException($"Invalid data: complex object found in leaf schema node {schemaNode.GetNodePath()}.");
-            }
-
-            return fhirLeafObject;
-        }
-
-        private JObject ProcessChoiceTypeObject(JToken fhirObject, FhirParquetSchemaNode schemaNode)
-        {
-            JObject choiceRootObject = new JObject();
-            if (schemaNode.IsLeaf)
-            {
-                choiceRootObject.Add(schemaNode.Name, ProcessLeafObject(fhirObject, schemaNode));
-            }
-            else
-            {
-                choiceRootObject.Add(schemaNode.Name, ProcessStructObject(fhirObject, schemaNode));
-            }
-
-            return choiceRootObject;
-        }
-
-        private MemoryStream ConvertJsonDataToStream(string schemaType, IEnumerable<JObject> inputData)
+        private MemoryStream TransformJsonDataToStream(string schemaType, IEnumerable<JObject> inputData)
         {
             var content = string.Join(
                 Environment.NewLine,
